requirements.txt
# 核心依赖
clickhouse-connect>=0.6.0
PyYAML>=6.0

# 开发依赖
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.900
